+++ swift-drawingapp/Model/Entities/Shape/Rectangle.swift
