+++ swift-drawingapp/Model/Entities/Shape/Selectable.swift
