+++ swift-drawingapp/Model/Entities/Style/Color.swift
