+++ swift-drawingapp/Model/Entities/Style/StyleApplying.swift
