+++ swift-drawingapp/View/Shape/ShapeView.swift
@@ -7,24 +7,6 @@
 
 import UIKit
 
-protocol ShapeViewRepresentable {
-    var cgPoints: [CGPoint] { get }
-    var cgFillColor: CGColor? { get }
-    var cgLineColor: CGColor? { get }
-}
-
-struct AnyShapeViewModel: ShapeViewRepresentable {
-    var cgPoints: [CGPoint]
-    var cgFillColor: CGColor?
-    var cgLineColor: CGColor?
-    
-    init(_ shapeViewModel: ShapeViewRepresentable) {
-        self.cgPoints = shapeViewModel.cgPoints
-        self.cgFillColor = shapeViewModel.cgFillColor
-        self.cgLineColor = shapeViewModel.cgLineColor
-    }
-}
-
 class ShapeView: UIView {
     
     lazy private var shape = {
